+++ src/ipred_prepare.rs
@@ -130,7 +130,7 @@ pub fn rav1d_prepare_intra_edges<BD: BitDepth>(
     angle: &mut c_int,
     tw: c_int,
     th: c_int,
-    filter_edge: c_int,
+    filter_edge: u8,
     topleft_out: &mut [BD::Pixel],
     topleft_out_offset: usize, // position of top-left sample in `topleft_out`
     bd: BD,
